+++ example/shieldwall/curve_basic.py
@@ -1,25 +1,28 @@
 import cadquery as cq
-from cqportal.shieldwall import CurveBasic, ArchShape
+from cqportal.shieldwall import CurveBasic, ShieldShape, Magnets
 
 
 wall_bp = CurveBasic()
-wall_bp.shape_bp = ArchShape()
+wall_bp.width = 20
+wall_bp.height = 25
+wall_bp.base_height = 5.6
 
-#90 degrees
+wall_bp.x_radius = 75
+wall_bp.y_radius = 75 #ellipse stretch
 wall_bp.angle = 270
-wall_bp.rotation_angle=0
+wall_bp.rotation_angle = 0
 
-#45 degrees
-#wall_bp.angle = 315
-#wall_bp.rotation_angle=-45
+wall_bp.render_magnets = True
+wall_bp.magnet_padding = 1
+wall_bp.magnet_padding_x = 2
 
-#60 degrees
-#wall_bp.angle = 300
-#wall_bp.rotation_angle=-30
+# blueprints
+wall_bp.shape_bp = ShieldShape()
+wall_bp.magnets_bp = Magnets()
 
 
 wall_bp.make()
 wall_ex = wall_bp.build()
 
 #show_object(wall_ex)
-cq.exporters.export(wall_ex, 'stl/shieldwall_curveBasic.stl')
+cq.exporters.export(wall_ex, 'stl/shieldwall_curve_basic.stl')