+++ example/shieldwall/hex_mesh.py
@@ -11,4 +11,4 @@ mesh_bp.make()
 mesh_ex = mesh_bp.build()
 
 #show_object(mesh_ex)
-cq.exporters.export(mesh_ex, 'stl/shieldwall_hexMesh.stl')
+cq.exporters.export(mesh_ex, 'stl/shieldwall_hex_mesh.stl')